Refactor model layout
