+++ src/registrar/models/__init__.py
@@ -1,6 +1,10 @@
 from auditlog.registry import auditlog  # type: ignore
 
-from .models import User, UserProfile, Contact, Website, DomainApplication
+from .contact import Contact
+from .domain_application import DomainApplication
+from .user_profile import UserProfile
+from .user import User
+from .website import Website
 
 __all__ = [
     "Contact",
